requirements.txt
# Core Django
Django==5.2.8
django-filter==25.2

# REST Framework
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1

# API Documentation
drf-spectacular==0.29.0

# CORS Headers
django-cors-headers==4.7.0

# Image Processing
Pillow==12.0.0

# Code Formatting (Development)
black==25.11.0

# Dependencies
asgiref==3.11.0
attrs==25.4.0
click==8.3.1
colorama==0.4.6
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
PyJWT==2.10.1
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
